Add carpools search results

+++ app/src/main/java/com/campusmov/uniride/domain/analytics/model/StudentRatingMetric.kt
@@ -2,7 +2,7 @@ package com.campusmov.uniride.domain.analytics.model
 
 data class StudentRatingMetric (
     val userId: String,
-    val totalRatings: Int,
-    val totalReviewsCount: Double,
+    val totalRatings: Double,
+    val totalReviewsCount: Int,
     val averageRating: Double
 )
